+++ abipy/core/func1d.py
@@ -502,14 +502,17 @@ class Function1D(object):
 
     def plotly_traces(self, fig, row=1, col=1, exchange_xy=False, xfactor=1, yfactor=1, *args, **kwargs):
         """
-        Helper function to plot self on axis ax.
+        Helper function to plot self with plotly.
+
         Args:
             fig: |plotly.graph_objects.Figure|.
-            row, col: only used when the fig has subplots, for specifying to add traces on which subplot.
-            exchange_xy: True to exchange the axis in the plot.
-            args: Positional arguments passed to 'plotly.graph_objects.Scatter'
+            row, col: only used when the fig has subplots, for specifying the subplot where traces should be added.
+            exchange_xy: True to exchange the x and y in the plot.
             xfactor, yfactor: xvalues and yvalues are multiplied by this factor before plotting.
-            kwargs: Keyword arguments passed to 'plotly.graph_objects.Scatter'. Accepts
+            args: Positional arguments passed to 'plotly.graph_objects.Scatter'
+            kwargs: Keyword arguments passed to 'plotly.graph_objects.Scatter'.
+                Accepts also Abipy specific kwargs:
+
         ==============  ===============================================================
         kwargs          Meaning
         ==============  ===============================================================
@@ -519,10 +522,7 @@ class Function1D(object):
                         "angle" to display the phase of the complex number in radians.
                         Options can be concatenated with "-"
         ==============  ===============================================================
-    # !!   Returns:
-    #        List of lines added.
         """
-
         import plotly.graph_objects as go
 
         if self.iscomplexobj:
@@ -530,7 +530,6 @@ class Function1D(object):
         else:
             cplx_mode = kwargs.pop("cplx_mode", "re")
 
-        #     lines = []
         for c in cplx_mode.lower().split("-"):
             xx, yy = self.mesh, data_from_cplx_mode(c, self.values)
             if xfactor != 1: xx = xx * xfactor
@@ -539,14 +538,11 @@ class Function1D(object):
             if exchange_xy:
                 xx, yy = yy, xx
 
-            #         lines.extend(ax.plot(xx, yy, *args, **kwargs))
             if row == 1 and col == 1:
                 fig.add_trace(go.Scatter(x=xx, y=yy, mode="lines", showlegend=False, *args, **kwargs))
 
             else:
                 fig.add_trace(go.Scatter(x=xx, y=yy, mode="lines", showlegend=False, *args, **kwargs), row=row, col=col)
-
-    #     return lines
 
     @add_fig_kwargs
     def plot(self, ax=None, **kwargs):